:sparkles: Apply displaying square on click
