+++ packages/web/src/components/MandalaChart/MandalaChart.tsx
@@ -2,6 +2,7 @@ import { MandalaChartView } from '@/constants/mandalaChart'
 import { useEntryContext } from '@/contexts/EntryContext'
 import { Dispatch, useEffect, useRef } from 'react'
 import { deepCopy } from '../../../utils/common'
+import { RecommendationItemProps } from '../Recommend/RecommendationItem'
 import FullViewMandalaChart from './FullViewMandalaChart'
 import SingleViewMandalaChart from './SingleViewMandalaChart'
 
@@ -10,6 +11,7 @@ interface Props {
   wholeGridValues: string[][]
   setWholeGridValues: Dispatch<React.SetStateAction<string[][]>>
   isAIModeOn: boolean
+  recommendationItems: RecommendationItemProps[]
 }
 
 const MandalaChart = ({
@@ -17,19 +19,27 @@ const MandalaChart = ({
   wholeGridValues,
   setWholeGridValues,
   isAIModeOn,
+  recommendationItems,
 }: Props) => {
   const { isMobile } = useEntryContext()
   const focusRef = useRef<HTMLDivElement>(null)
 
+  const selectedAIRecommendationItem = recommendationItems.find(
+    item => item.isClicked
+  )
   const handleGridValue = (
     gridIndex: number,
     squareIndex: number,
     newValue: string
   ) => {
     const newGridValues = deepCopy(wholeGridValues)
-    newGridValues[gridIndex][squareIndex] = newValue
+    newGridValues[gridIndex][squareIndex] = isAIModeOn
+      ? selectedAIRecommendationItem?.text
+      : newValue
     if (gridIndex === 4 && squareIndex !== 4) {
-      newGridValues[squareIndex][gridIndex] = newValue
+      newGridValues[squareIndex][gridIndex] = isAIModeOn
+        ? selectedAIRecommendationItem?.text
+        : newValue
     }
     setWholeGridValues(newGridValues)
   }
