+++ packages/web/src/components/Recommend/Recommendations.tsx
@@ -1,34 +1,12 @@
-import { useState } from 'react'
 import RecommendationItem, {
   RecommendationItemProps,
 } from './RecommendationItem'
 
 export interface Props {
-  items: { text: string }[]
+  recommendationItems: RecommendationItemProps[]
+  handleItemClick: (id: number) => () => void
 }
-const Recommendations = ({ items }: Props) => {
-  const [recommendationItems, setRecommendationItems] = useState<
-    RecommendationItemProps[]
-  >(
-    items.map((item, idx) => ({
-      id: idx,
-      text: item.text,
-      isClicked: false,
-    }))
-  )
-
-  const handleItemClick = (id: number) => () => {
-    const previousClickedItem = recommendationItems.find(item => item.isClicked)
-    const updatedItems = recommendationItems.map(item => ({
-      ...item,
-      isClicked: item.id === id ? !item.isClicked : false,
-    }))
-    if (previousClickedItem?.isClicked) {
-      updatedItems[previousClickedItem.id].isClicked = false
-    }
-    setRecommendationItems(updatedItems)
-  }
-
+const Recommendations = ({ recommendationItems, handleItemClick }: Props) => {
   return (
     <div className="fixed bottom-10 left-10 right-10 bg-white shadow-lg p-4 rounded-lg">
       <div className="grid grid-cols-2 md:grid-cols-4 gap-4">
